Time: 43 ms (5.28%), Space: 28.1 MB (5.80%) - LeetHub

+++ 968-binary-tree-cameras/968-binary-tree-cameras.cpp
@@ -1,5 +1,5 @@
 class Solution {
-    unordered_map<TreeNode*, int> dp[2][2];
+    unordered_map<TreeNode*, int64_t> dp[2][2];
     
     int64_t solve(TreeNode* root, bool mandatory, bool covered) {
